sources/app/requirements.txt
nltk
mlflow
pandas
pymongo
xgboost
scikit-learn
python-dotenv
azure-storage-blob